pyproject.toml
[tool.poetry]
name = "django-project-template"
version = "0.0.0"
description = "Django project template to speed up the creation of django based applications"
authors = ["malathair <malathair@spectregate.net>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
django-environ = "^0.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
